Update metadata.json
Added current default Debian 12 gnome-shell version 43. I tested this extension on gnome-shell 43.9 and it works great! Unfortunately, this extension can't be enabled in Debian 12 without explicitly specifying the gnome-shell version in metadata.js
This commit fixes that problem and makes this extension work out of the box on Debian systems